+++ packages/core/src/indexing/local-index-api.ts
@@ -26,14 +26,6 @@ export class LocalIndexApi implements IndexApi {
     private readonly logger: DiagnosticsLogger,
     networkName: Networks
   ) {
-    // the '?' is here, because we're passing `unknown` as the indexingConfig in tests. refactor the tests maybe?
-    if (indexingConfig?.models !== undefined) {
-      // TODO: Update the log once admin API is implemented
-      logger.warn(`
-      Passing the list of Model StreamIDs to be indexed the config file's 'indexing.models' field is deprecated and will soon be removed as an option.
-      Instead, you will be configuring your index via the @composedb/cli commands (see documentation at https://composedb.js.org).
-      `)
-    }
     this.databaseIndexApi = makeIndexApi(indexingConfig, networkName, logger)
   }
 
@@ -111,14 +103,16 @@ export class LocalIndexApi implements IndexApi {
   async indexModels(models: Array<StreamID>): Promise<void> {
     // TODO: Load model StreamIDs, extract relations and pass those arguments down into the
     //  DatabaseIndexAPI so the necessary columns get built for the relations
-    await this.databaseIndexApi?.indexModels(models)
+    const indexModelArgs = []
+    for (const model of models) {
+      const args = { model }
+      indexModelArgs.push(args)
+    }
+    await this.databaseIndexApi?.indexModels(indexModelArgs)
   }
 
   async init(): Promise<void> {
-    if (this.indexingConfig.models) {
-      return this.indexModels(this.indexingConfig.models)
-    }
-    return
+    return this.indexModels(this.indexingConfig.models)
   }
 
   async close(): Promise<void> {