+++ packages/core/src/indexing/postgres/__tests__/insertion-order.test.ts
@@ -30,7 +30,11 @@ beforeEach(async () => {
     connection: process.env.DATABASE_URL,
   })
   const indexAPI = new PostgresIndexApi(dbConnection, true, logger)
-  await indexAPI.indexModels(MODELS_TO_INDEX)
+  await indexAPI.indexModels(
+    MODELS_TO_INDEX.map((model) => {
+      return { model }
+    })
+  )
   order = new InsertionOrder(dbConnection)
   // Rows in insertion-order.fixture.csv are in insertion order.
   // The responses in the tests below are ok if they are in the same order as in the CSV.