+++ packages/core/src/indexing/sqlite/migrations/mid-schema-verfication.ts
@@ -1,55 +1,65 @@
 /**
- * Valid SQLite table structure for mid tables
- * Used to verify table integrity during node startup
+ * Expected SQLite structure for columns for relations.
+ * Used to verify table integrity during node startup and after indexing a new model.
  */
+export const RELATION_COLUMN_STRUCTURE = {
+  type: 'varchar',
+  maxLength: '1024',
+  nullable: false,
+  defaultValue: null,
+}
 
-export const validTableStructure = {
+/**
+ * Valid SQLite table structure for mid tables
+ * Used to verify table integrity during node startup and after indexing a new model.
+ */
+export const COMMON_TABLE_STRUCTURE = {
   stream_id: {
     type: 'varchar',
     maxLength: '1024',
     nullable: false,
-    defaultValue: null
+    defaultValue: null,
   },
   controller_did: {
     type: 'varchar',
     maxLength: '1024',
     nullable: false,
-    defaultValue: null
+    defaultValue: null,
   },
   stream_content: {
     type: 'varchar',
     maxLength: '255',
     nullable: false,
-    defaultValue: null
+    defaultValue: null,
   },
   tip: {
     type: 'varchar',
     maxLength: '255',
     nullable: false,
-    defaultValue: null
+    defaultValue: null,
   },
   last_anchored_at: {
     type: 'integer',
     maxLength: null,
     nullable: true,
-    defaultValue: null
+    defaultValue: null,
   },
   first_anchored_at: {
     type: 'integer',
     maxLength: null,
     nullable: true,
-    defaultValue: null
+    defaultValue: null,
   },
   created_at: {
     type: 'integer',
     maxLength: null,
     nullable: false,
-    defaultValue: null
+    defaultValue: null,
   },
   updated_at: {
     type: 'integer',
     maxLength: null,
     nullable: false,
-    defaultValue: null
-  }
+    defaultValue: null,
+  },
 }