+++ packages/core/src/state-management/repository.ts
@@ -39,6 +39,15 @@ export type RepositoryDependencies = {
 
 const DEFAULT_LOAD_OPTS = { sync: SyncOptions.PREFER_CACHE, syncTimeoutSeconds: 3 }
 
+/**
+ * Indicate if the stream should be indexed.
+ */
+function shouldIndex(state$: RunningState, index: LocalIndexApi): boolean {
+  const model = state$.state?.metadata?.model
+  if (!model) return false
+  return index.shouldIndexStream(model)
+}
+
 export class Repository {
   /**
    * Serialize loading operations per streamId.
@@ -283,8 +292,8 @@ export class Repository {
     await this.handlePinOpts(state$, opts)
   }
 
-  async handlePinOpts(state$: RunningState, opts: PinningOpts) {
-    if (opts.pin) {
+  async handlePinOpts(state$: RunningState, opts: PinningOpts): Promise<void> {
+    if (opts.pin || (opts.pin === undefined && shouldIndex(state$, this.index))) {
       await this.pin(state$)
     } else if (opts.pin === false) {
       await this.unpin(state$)
@@ -337,9 +346,18 @@ export class Repository {
   }
 
   async unpin(state$: RunningState, opts?: PublishOpts): Promise<void> {
+    if (shouldIndex(state$, this.index)) {
+      throw new Error(
+        `Cannot unpin actively indexed stream (${state$.id.toString()}) with model: ${
+          state$.state.metadata.model
+        }`
+      )
+    }
+
     if (opts?.publish) {
       this.stateManager.publishTip(state$)
     }
+
     return this.#deps.pinStore.rm(state$)
   }