+++ packages/ipfs-topology/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.1](/compare/@ceramicnetwork/ipfs-topology@2.2.1-rc.3...@ceramicnetwork/ipfs-topology@2.2.1) (2022-09-21)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
+## [2.2.1-rc.3](/compare/@ceramicnetwork/ipfs-topology@2.2.1-rc.2...@ceramicnetwork/ipfs-topology@2.2.1-rc.3) (2022-09-20)
+
+**Note:** Version bump only for package @ceramicnetwork/ipfs-topology
+
+
+
+
+
 ## [2.2.1-rc.2](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/ipfs-topology@2.2.1-rc.1...@ceramicnetwork/ipfs-topology@2.2.1-rc.2) (2022-09-15)
 
 **Note:** Version bump only for package @ceramicnetwork/ipfs-topology