+++ packages/stream-model-instance-handler/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.7.0](/compare/@ceramicnetwork/stream-model-instance-handler@0.7.0-rc.0...@ceramicnetwork/stream-model-instance-handler@0.7.0) (2022-09-21)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler
+
+
+
+
+
+# [0.7.0-rc.0](/compare/@ceramicnetwork/stream-model-instance-handler@0.6.1-rc.2...@ceramicnetwork/stream-model-instance-handler@0.7.0-rc.0) (2022-09-20)
+
+
+### Features
+
+* **common:** Update type definitions to support simple relations (#2421) a4c4ce3, closes #2421
+
+
+
+
+
 ## [0.6.1-rc.2](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-model-instance-handler@0.6.1-rc.1...@ceramicnetwork/stream-model-instance-handler@0.6.1-rc.2) (2022-09-15)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance-handler