+++ packages/stream-model-instance/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.2](/compare/@ceramicnetwork/stream-model-instance@0.4.2-rc.3...@ceramicnetwork/stream-model-instance@0.4.2) (2022-09-21)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance
+
+
+
+
+
+## [0.4.2-rc.3](/compare/@ceramicnetwork/stream-model-instance@0.4.2-rc.2...@ceramicnetwork/stream-model-instance@0.4.2-rc.3) (2022-09-20)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-model-instance
+
+
+
+
+
 ## [0.4.2-rc.2](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-model-instance@0.4.2-rc.1...@ceramicnetwork/stream-model-instance@0.4.2-rc.2) (2022-09-15)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-model-instance