+++ packages/stream-tile-handler/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.2.8](/compare/@ceramicnetwork/stream-tile-handler@2.2.8-rc.3...@ceramicnetwork/stream-tile-handler@2.2.8) (2022-09-21)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
+## [2.2.8-rc.3](/compare/@ceramicnetwork/stream-tile-handler@2.2.8-rc.2...@ceramicnetwork/stream-tile-handler@2.2.8-rc.3) (2022-09-20)
+
+**Note:** Version bump only for package @ceramicnetwork/stream-tile-handler
+
+
+
+
+
 ## [2.2.8-rc.2](https://github.com/ceramicnetwork/js-ceramic/compare/@ceramicnetwork/stream-tile-handler@2.2.8-rc.1...@ceramicnetwork/stream-tile-handler@2.2.8-rc.2) (2022-09-15)
 
 **Note:** Version bump only for package @ceramicnetwork/stream-tile-handler